Refactoring artwork fetching logic for readability.
