+++ Shared/Core/Models and Services/Artwork Service/ArtworkFetcher.swift
@@ -1,0 +1,9 @@
+//
+//  ArtworkFetcher.swift
+//  Core
+//
+//  Created by Jake Bromberg on 4/2/23.
+//  Copyright © 2023 WXYC. All rights reserved.
+//
+
+import Foundation
